+++ src/app/componant/dentist-point/edit-template/edit-template.component.html
@@ -1,5 +1,4 @@
 <div class="col-md-12 ibox">
-
   <form name="form" [formGroup]="form">
     <div class="ibox-head">
       <div class="ibox-title">Edit Template</div>
@@ -231,7 +230,7 @@
               <td>{{ createMedicine.drugStrength }}</td>
               <td>{{ createMedicine.drugDose }}</td>
               <td>{{ createMedicine.drugDuration }}</td>
-              <button class="btn btn-outline-danger mt-1"  type="button">Delete</button>
+              <button class="btn btn-outline-danger mt-1" type="button">Delete</button>
             </tr>
             </tbody>
           </table>
