+++ src/app/service/template.service.ts
@@ -1,66 +1,14 @@
 import {Injectable} from '@angular/core';
-import {environment} from '../../environments/environment';
-import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {AuthService} from './auth.service';
 import {Observable} from 'rxjs';
-import {EmptyObservable} from "rxjs-compat/observable/EmptyObservable";
-import {CreateDrug} from "../model/create-medicine";
+import {CreateDrug} from '../model/create-medicine';
 
 @Injectable()
-export class TemplateService {
+export abstract class TemplateService {
+  abstract createTemplate(tempName: string, complain: string, parameters: string, remarks: string, dentalHistory: string,
+                          vaccinationHistory: string, investigation: string, radiological: string, planning: string,
+                          prescriptionList: Array<CreateDrug>): Observable<any>;
 
-  private enableMock: boolean = environment.chumbok.enableMock;
-  private callThroughGateway: boolean = environment.chumbok.apiCallThroughGateway;
-
-  private orgListEndpoint: string = this.callThroughGateway ?
-    environment.chumbok.apiBaseEndpoint + '/uaa/orgs' : environment.chumbok.apiBaseEndpoint + '/orgs';
-
-  private serviceBaseEndpoint: string = this.callThroughGateway ?
-    environment.chumbok.apiBaseEndpoint + '/uaa' : environment.chumbok.apiBaseEndpoint;
-
-  private httpOptions = {
-    headers: new HttpHeaders({'Authorization': 'Bearer ' + this.authService.getAuthToken()})
-  };
-
-  private authHeader = new HttpHeaders({'Authorization': 'Bearer ' + this.authService.getAuthToken()});
-
-  constructor(private authService: AuthService, private http: HttpClient) {
-  }
-
-
-  public createTemplate(tempName: string, complain: string, parameters: string, remarks: string, dentalHistory: string,
-                        vaccinationHistory: string, investigation: string, radiological: string, planning: string,
-                        prescriptionList: Array<CreateDrug>): Observable<any> {
-    console.log("template", tempName);
-    console.log(complain);
-    console.log(parameters);
-    console.log(remarks);
-    console.log(dentalHistory);
-    console.log(vaccinationHistory);
-    console.log(investigation);
-    console.log(radiological);
-    console.log(prescriptionList)
-
-    return new EmptyObservable<Response>();
-
-  }
-
-  public editTemplate(tempId: string, tempName: string, complain: string, parameters: string,
-                      remarks: string, dentalHistory: string, vaccinationHistory: string, investigation: string,
-                      radiological: string, planning: string, prescriptionList: Array<CreateDrug>): Observable<any> {
-    console.log(tempId);
-    console.log("template name:", tempName);
-    console.log(complain);
-    console.log(parameters);
-    console.log(remarks);
-    console.log(dentalHistory);
-    console.log(vaccinationHistory);
-    console.log(investigation);
-    console.log(radiological);
-    console.log(prescriptionList)
-
-    return new EmptyObservable<Response>();
-
-  }
-
+  abstract editTemplate(tempId: string, tempName: string, complain: string, parameters: string,
+                        remarks: string, dentalHistory: string, vaccinationHistory: string, investigation: string,
+                        radiological: string, planning: string, prescriptionList: Array<CreateDrug>): Observable<any>;
 }
